\name{HGB_0013.fasta}
\alias{HGB_0013.fasta}
\docType{data}
\title{
FASTA alignment corresponding to OTU "HGB_0013_GXJPMPL01A3OQX"
}
\description{
FASTA alignment corresponding to OTU "HGB_0013_GXJPMPL01A3OQX".
}
\usage{data(HGB_0013.fasta)}
\format{
The format is a List of 2
 $ Seqs.names
 $ Seqs.only
}
\details{
}
\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
}
\references{
%%  ~~ possibly secondary sources and usages ~~
}
\examples{
data(HGB_0013.fasta)
str(HGB_0013.fasta)
}
\keyword{datasets}
