\name{otu2ot-package}
\alias{otu2ot-package}
\alias{otu2ot}
\docType{package}
\title{
The package implements the oligotyping procedure in R.
}
\description{
The package implements the oligotyping procedure in R.
}
\details{
\tabular{ll}{
Package: \tab otu2ot\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2014-07-31\cr
License: \tab GPL-2\cr
}
}
\author{
Alban Ramette
<ramette.alban@gmail.com>
}
\references{
}



